+++ pennylane_lightning/core/src/simulators/lightning_gpu/observables/ObservablesGPU.cpp
@@ -29,5 +29,5 @@ template class Observables::TensorProdObs<StateVectorCudaManaged<double>>;
 template class Observables::Hamiltonian<StateVectorCudaManaged<float>>;
 template class Observables::Hamiltonian<StateVectorCudaManaged<double>>;
 
-template class Observables::SparseHermitianObs<StateVectorCudaManaged<float>>;
-template class Observables::SparseHermitianObs<StateVectorCudaManaged<double>>;
+template class Observables::SparseHamiltonian<StateVectorCudaManaged<float>>;
+template class Observables::SparseHamiltonian<StateVectorCudaManaged<double>>;