+++ pennylane_lightning/core/src/simulators/lightning_kokkos/observables/ObservablesKokkos.cpp
@@ -29,5 +29,5 @@ template class Observables::TensorProdObs<StateVectorKokkos<double>>;
 template class Observables::Hamiltonian<StateVectorKokkos<float>>;
 template class Observables::Hamiltonian<StateVectorKokkos<double>>;
 
-template class Observables::SparseHermitianObs<StateVectorKokkos<float>>;
-template class Observables::SparseHermitianObs<StateVectorKokkos<double>>;
+template class Observables::SparseHamiltonian<StateVectorKokkos<float>>;
+template class Observables::SparseHamiltonian<StateVectorKokkos<double>>;