+++ pennylane_lightning/core/src/simulators/lightning_qubit/bindings/LQubitBindings.hpp
@@ -360,9 +360,9 @@ void registerBackendSpecificObservables([[maybe_unused]] py::module_ &m) {
 
     std::string class_name;
 
-    class_name = "SparseHermitianObsC" + bitsize;
-    py::class_<SparseHermitianObs<StateVectorT>,
-               std::shared_ptr<SparseHermitianObs<StateVectorT>>,
+    class_name = "SparseHamiltonianC" + bitsize;
+    py::class_<SparseHamiltonian<StateVectorT>,
+               std::shared_ptr<SparseHamiltonian<StateVectorT>>,
                Observable<StateVectorT>>(m, class_name.c_str(),
                                          py::module_local())
         .def(py::init([](const np_arr_c &data,
@@ -373,22 +373,21 @@ void registerBackendSpecificObservables([[maybe_unused]] py::module_ &m) {
             const py::buffer_info buffer_data = data.request();
             const auto *data_ptr = static_cast<ComplexT *>(buffer_data.ptr);
 
-            return SparseHermitianObs<StateVectorT>{
+            return SparseHamiltonian<StateVectorT>{
                 std::vector<ComplexT>({data_ptr, data_ptr + data.size()}),
                 indices, indptr, wires};
         }))
-        .def("__repr__", &SparseHermitianObs<StateVectorT>::getObsName)
-        .def("get_wires", &SparseHermitianObs<StateVectorT>::getWires,
+        .def("__repr__", &SparseHamiltonian<StateVectorT>::getObsName)
+        .def("get_wires", &SparseHamiltonian<StateVectorT>::getWires,
              "Get wires of observables")
         .def(
             "__eq__",
-            [](const SparseHermitianObs<StateVectorT> &self,
+            [](const SparseHamiltonian<StateVectorT> &self,
                py::handle other) -> bool {
-                if (!py::isinstance<SparseHermitianObs<StateVectorT>>(other)) {
+                if (!py::isinstance<SparseHamiltonian<StateVectorT>>(other)) {
                     return false;
                 }
-                auto other_cast =
-                    other.cast<SparseHermitianObs<StateVectorT>>();
+                auto other_cast = other.cast<SparseHamiltonian<StateVectorT>>();
                 return self == other_cast;
             },
             "Compare two observables");